Use executeOnWindow
